+++ components/semotus/src/helpers/ProcessCall.ts
@@ -53,7 +53,7 @@ async function retryCall(payload: ProcessCallPayload) {
  * @returns  unknown
  */
 function preCallHook(payload: ProcessCallPayload, forceupdate?: boolean): boolean {
-    const {semotus, remoteCall, callContext, session, subscriptionId, remoteCallId, restoreSessionCallback} = payload;
+    const {semotus, remoteCall, session, callContext, HTTPObjs} = payload;
     semotus.logger.info(
         {
             component: 'semotus',
@@ -74,12 +74,20 @@ function preCallHook(payload: ProcessCallPayload, forceupdate?: boolean): boolea
             changes[semotus.__dictionary__[objId.replace(/[^-]*-/, '').replace(/-.*/, '')].__name__] = true;
         }
 
+        let remoteObject = session.objects[remoteCall.id];
+
+        let isPublic = semotus.role === 'server' && remoteObject[remoteCall.name].remotePublic;
+
         return semotus.controller.preServerCall.call(
             semotus.controller,
             remoteCall.changes.length > 2,
             changes,
             callContext,
-            forceupdate
+            forceupdate,
+            remoteCall.name,
+            remoteCall,
+            isPublic,
+            HTTPObjs
         );
     } else {
         return true;
@@ -223,11 +231,11 @@ async function callIfValid(payload: ProcessCallPayload, isValid: boolean) {
  * @returns
  */
 async function postCallHook(payload: ProcessCallPayload, returnValue) {
-    const {semotus, remoteCall, callContext} = payload;
+    const {semotus, remoteCall, callContext, HTTPObjs} = payload;
 
     if (semotus.controller && semotus.controller.postServerCall) {
         const hasChanges: boolean = remoteCall.changes.length > 2;
-        await semotus.controller.postServerCall.call(semotus.controller, hasChanges, callContext, semotus.changeString);
+        await semotus.controller.postServerCall.call(semotus.controller, hasChanges, callContext, semotus.changeString, HTTPObjs);
     }
     return returnValue;
 }